Show View Finish

+++ app/src/main/java/tw/pitawanpor/twschool/MainActivity.java
@@ -127,7 +127,7 @@ public class MainActivity extends AppCompatActivity {
                         case 1:
                             //Student
 
-                            if (Integer.parseInt(loginStrings[5]) == 0) {
+                            if (Double.parseDouble(loginStrings[5]) == 0) {
                                 //Non Setup position Home
                                 Intent intent = new Intent(MainActivity.this, EditStudent.class);
                                 intent.putExtra("Login", loginStrings);
